+++ src/Collection/ObjectCollection.php
@@ -56,19 +56,6 @@
 
 
     /**
-     * Add one item to begin of collection
-     * This item is accessible via `$collection->getFirst();`
-     *
-     * @param ElementFinder $item
-     * @return self
-     */
-    public function prepend(ElementFinder $item) : self {
-      array_unshift($this->items, $item);
-      return $this;
-    }
-
-
-    /**
      * Add one item to the end of collection
      * This item is accessible via `$collection->getLast();`
      *
@@ -77,26 +64,6 @@
      */
     public function append(ElementFinder $item) : self {
       $this->items[] = $item;
-      return $this;
-    }
-
-
-    /**
-     * @param int $index
-     * @param ElementFinder[] $items
-     * @return self
-     * @throws \Exception
-     */
-    public function addAfter(int $index, array $items) : self {
-
-      foreach ($items as $item) {
-        $this->validateType($item);
-      }
-
-      $offset = $index + 1;
-      $firstPart = array_slice($this->items, 0, $offset);
-      $secondPart = array_slice($this->items, $offset);
-      $this->items = array_merge($firstPart, $items, $secondPart);
       return $this;
     }
 
@@ -113,36 +80,6 @@
         $this->validateType($item);
       }
       $this->items = $items;
-      return $this;
-    }
-
-
-    /**
-     * Remove part of items from collection
-     * Works as array_slice
-     *
-     * @param int $offset
-     * @param int|null $length
-     * @return self
-     */
-    public function slice(int $offset, int $length = null) : self {
-      $this->items = array_slice($this->items, $offset, $length);
-      return $this;
-    }
-
-
-    /**
-     * Take part of items and return new collection
-     * Works as array_slice
-     * At this point items in 2 collection is same
-     *
-     * @param int $offset
-     * @param int|null $length
-     * @return self
-     */
-    public function extractItems(int $offset, int $length = null) : self {
-      $items = array_slice($this->items, $offset, $length);
-      $this->setItems($items);
       return $this;
     }
 
@@ -170,32 +107,6 @@
         return null;
       }
       return reset($this->items);
-    }
-
-
-    /**
-     * Return next item from current
-     * Also can return item with position from current + $step
-     *
-     * @param int $step
-     * @return null|ElementFinder
-     */
-    public function getNext(int $step = 1) {
-      $position = ($this->position + $step);
-      return $this->items[$position] ?? null;
-    }
-
-
-    /**
-     * Return previous item
-     * Also can return previous from current position + $step
-     *
-     * @param int $step
-     * @return null|ElementFinder
-     */
-    public function getPrevious(int $step = 1) {
-      $position = ($this->position - $step);
-      return $this->items[$position] ?? null;
     }